grs/requirement.txt
asn1crypto==0.24.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cryptography==2.2.2
Django==1.9
django-mysql==2.4.1
enum34==1.1.6
gitdb2==2.0.3
GitPython==2.1.9
gunicorn==19.9.0
idna==2.6
ipaddress==1.0.22
mysql==0.0.1
MySQL-python==1.2.5
pycparser==2.18
pyOpenSSL==18.0.0
PyYAML==3.13
requests==2.18.4
six==1.11.0
smmap2==2.0.3
urllib3==1.22
